+++ proyectoperfulandia/src/main/java/com/example/proyectoperfulandia/controller/EmpleadoController.java
@@ -1,5 +1,6 @@
 package com.example.proyectoperfulandia.controller;
 
+import com.example.proyectoperfulandia.assembler.EmpleadoModelAssembler;
 import com.example.proyectoperfulandia.model.Empleado;
 import com.example.proyectoperfulandia.services.EmpleadoService;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -10,7 +11,13 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.hateoas.CollectionModel;
+import org.springframework.hateoas.EntityModel;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/empleados")
@@ -20,6 +27,9 @@ public class EmpleadoController {
     @Autowired
     EmpleadoService empleadoService;
 
+    @Autowired
+    EmpleadoModelAssembler assembler;
+
     // Obtener solo los usuarios con el rol EMPLEADO
     @GetMapping
     @Operation(summary = "Obtener listado de todos los empleados")
@@ -27,8 +37,12 @@ public class EmpleadoController {
             @ApiResponse(responseCode = "200", description = "Listado de usuarios generado exitosamente."),
             @ApiResponse(responseCode = "404", description = "Listado no encontrado.")
     })
-    public String getEmpleados(){
-        return empleadoService.getEmpleados();
+    public ResponseEntity<CollectionModel<EntityModel<Empleado>>> getEmpleados() {
+        List<Empleado> empleado = empleadoService.getEmpleados();
+        if (empleado.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(assembler.toCollectionModel(empleado), HttpStatus.OK);
     }
 
     // Añadir un usuario EMPLEADO
@@ -39,8 +53,15 @@ public class EmpleadoController {
                     schema = @Schema(implementation = Empleado.class))),
             @ApiResponse(responseCode = "204", description = "No hay contenido en la solicitud.")
     })
-    public String addEmpleado(@RequestBody Empleado empleado){
-        return empleadoService.addEmpleado(empleado);
+    public ResponseEntity<EntityModel<Empleado>> addEmpleado(
+            @Parameter(description = "Datos del usuario EMPLEADO a agregar", required = true)
+            @RequestBody Empleado empleado) {
+        try {
+            empleadoService.addEmpleado(empleado);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     // Obtener un usuario EMPLEADO mediante ID
@@ -50,9 +71,12 @@ public class EmpleadoController {
             @ApiResponse(responseCode = "200", description = "Usuario encontrado exitosamente."),
             @ApiResponse(responseCode = "404", description = "Usuario no encontrado.")
     })
-    @Parameter(description = "ID del empleado a obtener", required = true, example = "1")
-    public String getEmpleado(@PathVariable int id){
-        return empleadoService.getEmpleadoID(id);
+    public ResponseEntity<EntityModel<Empleado>> getEmpleado(
+            @Parameter(description = "ID del usuario EMPLEADO a obtener", required = true, example = "1")
+            @PathVariable int id){
+        return empleadoService.getEmpleado(id)
+                .map(empleado -> new ResponseEntity<>(assembler.toModel(empleado), HttpStatus.OK))
+                .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     // Eliminar un usuario EMPLEADO mediante ID
@@ -63,8 +87,14 @@ public class EmpleadoController {
             @ApiResponse(responseCode = "404", description = "Usuario no encontrado.")
     })
     @Parameter(description = "ID del empleado a eliminar", required = true, example = "1")
-    public String removeEmpleado(@PathVariable int id){
-        return empleadoService.removeEmpleado(id);
+    public ResponseEntity<?> removeEmpleado(
+            @Parameter(description = "ID del usuario EMPLEADO a eliminar", required = true, example = "1")
+            @PathVariable int id){
+        if (!empleadoService.existsById(id)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        empleadoService.removeEmpleado(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     // Actualizar un usuario EMPLEADO mediante ID
@@ -75,9 +105,17 @@ public class EmpleadoController {
                     schema = @Schema(implementation = Empleado.class))),
             @ApiResponse(responseCode = "204", description = "No hay contenido en la solicitud.")
     })
-    @Parameter(description = "ID del empleado a actualizar", required = true, example = "1")
-    public String updateEmpleado(@PathVariable int id, @RequestBody Empleado empleado){
-        return empleadoService.updateEmpleado(id, empleado);
+    public ResponseEntity<EntityModel<Empleado>> updateEmpleado(
+            @Parameter(description = "ID del usuario ADMIN a actualizar", required = true, example = "1")
+            @PathVariable int id,
+            @Parameter(description = "Datos del usuario ADMIN a actualizar", required = true)
+            @RequestBody Empleado empleado) {
+        if (!empleadoService.existsById(id)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        empleado.setId(id);
+        empleadoService.updateEmpleado(id, empleado);
+        return new ResponseEntity<>(assembler.toModel(empleado), HttpStatus.OK);
     }
 
 }
